+++ teste/Spotify/src/app/app.component.html
@@ -1,3 +1,2 @@
-<app-nav-bar></app-nav-bar>
-<app-show-case></app-show-case>
-<app-footer></app-footer>
+<app-nav-bar> </app-nav-bar>
+<router-outlet> </router-outlet>
