+++ teste/Spotify/src/app/faq/faq.component.css
@@ -21,11 +21,6 @@ details div.content {
   padding: 0px 24px 24px 24px;
 }
 
-details[open] summary:after {
-  color: #9e9e9e;
-  content: "expand_less";
-  font-family: "Material Icons";
-}
 details[open] {
   margin-top: 16px;
   margin-bottom: 16px;
@@ -77,20 +72,6 @@ summary ul li span {
 }
 summary::-webkit-details-marker {
   display: none;
-}
-summary::after {
-  content: "expand_more";
-  font-family: "Material Icons";
-  color: #9e9e9e;
-  font-size: 1.5em;
-  padding: 0;
-  text-align: center;
-  margin-top: 0;
-  position: absolute;
-  top: calc(50% - 12px);
-  bottom: 0;
-  right: 0;
-  width: 6%;
 }
 
 .titleValue {
