--- teste/Spotify/src/app/footbar/footbar.component.css
@@ -1,4 +1,0 @@
-footer p {
-  color: #fff;
-  text-align: center;
-}
