--- teste/Spotify/src/app/footbar/footbar.component.html
@@ -1,5 +1,0 @@
-<footer>
-  <p>
-    Copyright&copy; {{ getTodayDate() }}, Spotify. Todos os direitos reservados
-  </p>
-</footer>
