+++ teste/Spotify/src/app/footer/footer.component.html
@@ -1,75 +1,106 @@
-<footer>   <!--RODAPÉ-->
-    <div class = "wrapper"> <!--Para modificar largura e altura-->  
+<footer>
+  <!--RODAPÉ-->
+  <div class="wrapper">
+    <!--Para modificar largura e altura-->
 
-        <div class = "footer-box">  <!--Colunas-->
-            <div class = "company-footer">
-                <a href="index.html">
-                <img src="assets/Spotify-logo.png" width="170" height="50" >
-                </a>
-            </div>
-        </div>
-
-        <div class = "footer-box">
-            <div     class = "articles-footer">
-            <h2>PlayLists Do Momento</h2>
-            <ul class = "footer-list footer-articles-list">
-                
-                <li><a href="https://open.spotify.com/playlist/37i9dQZF1DWWQRwui0ExPn" target="_blank">Lo-Fi Beats 2020
-                <img src="assets/playlist.png" width="20" height="20"></a></li>
-                <li><a href="https://open.spotify.com/playlist/5sTHqyG2DAwmTCopHXHRdz" target="_blank">Pop-Mordern 2020 </a>
-                    <img src="assets/playlist.png" width="20" height="20" ></li>
-                <li><a href="https://open.spotify.com/playlist/0U6RJkZ9aofC9g9dR72hlz" target="_blank">Eletronica-List 2020 </a>
-                    <img src="assets/playlist.png" width="20" height="20"></li>
-
-            </ul>
-        </div>
-        </div>
-
-        <div class = "footer-box">
-            <h2>Redes Sociais</h2>
-            <ul class="footer-list">
-                <div class="imageFacebook">    
-                <li>
-                   
-                    <a href="https://www.facebook.com/SpotifyBrasil/" target="_blank">
-                        <img src="assets/facebook.png" alt="Qries" width="50" height="50">
-                        <i class="fab fa-facebook-square"></i>
-                        
-                    </a>
-
-                </li>
-            </div>
-            
-            <div class="imageInstagram">
-                <li>
-                    <a href="https://www.instagram.com/spotifybrasil/?hl=pt-br" target="_blank">
-                        <img src="assets/instagram.png" alt="Qries" width="50" height="50">
-                        <i class="fab fa-instagram"></i>
-                        
-                    </a>
-
-                </li>
-            </div>
-
-            <div class="imageTwitter">
-                <li>
-                    <a href="https://twitter.com/SpotifyBR?ref_src=twsrc%5Egoogle%7Ctwcamp%5Eserp%7Ctwgr%5Eauthor" target="_blank">
-                        <img src="assets/twitter.png" alt="Qries" width="50" height="50">
-                        <i class="fab fa-twitter"></i>
-                        
-                    </a>
-
-                </li>
-            </div>
-
-                
-            </ul>
-        </div>
-
+    <div class="footer-box">
+      <!--Colunas-->
+      <div class="company-footer">
+        <a href="index.html">
+          <img src="assets/Spotify-logo.png" width="170" height="50" />
+        </a>
+      </div>
     </div>
-<div class = "footer-bottom">
+
+    <div class="footer-box">
+      <div class="articles-footer">
+        <h2>PlayLists Do Momento</h2>
+        <ul class="footer-list footer-articles-list">
+          <li>
+            <a
+              href="https://open.spotify.com/playlist/37i9dQZF1DWWQRwui0ExPn"
+              target="_blank"
+              >Lo-Fi Beats 2020
+              <img src="assets/playlist.png" width="20" height="20"
+            /></a>
+          </li>
+          <li>
+            <a
+              href="https://open.spotify.com/playlist/5sTHqyG2DAwmTCopHXHRdz"
+              target="_blank"
+              >Pop-Mordern 2020
+            </a>
+            <img src="assets/playlist.png" width="20" height="20" />
+          </li>
+
+          <li>
+            <a
+              href="https://open.spotify.com/playlist/0U6RJkZ9aofC9g9dR72hlz"
+              target="_blank"
+              >Eletronica-List 2020
+            </a>
+            <img src="assets/playlist.png" width="20" height="20" />
+          </li>
+        </ul>
+      </div>
+    </div>
+
+    <div class="footer-box">
+      <h2>Redes Sociais</h2>
+      <ul class="footer-list">
+        <div class="imageFacebook">
+          <li>
+            <a href="https://www.facebook.com/SpotifyBrasil/" target="_blank">
+              <img
+                src="assets/facebook.png"
+                alt="Qries"
+                width="50"
+                height="50"
+              />
+              <i class="fab fa-facebook-square"></i>
+            </a>
+          </li>
+        </div>
+
+        <div class="imageInstagram">
+          <li>
+            <a
+              href="https://www.instagram.com/spotifybrasil/?hl=pt-br"
+              target="_blank"
+            >
+              <img
+                src="assets/instagram.png"
+                alt="Qries"
+                width="50"
+                height="50"
+              />
+              <i class="fab fa-instagram"></i>
+            </a>
+          </li>
+        </div>
+
+        <div class="imageTwitter">
+          <li>
+            <a
+              href="https://twitter.com/SpotifyBR?ref_src=twsrc%5Egoogle%7Ctwcamp%5Eserp%7Ctwgr%5Eauthor"
+              target="_blank"
+            >
+              <img
+                src="assets/twitter.png"
+                alt="Qries"
+                width="50"
+                height="50"
+              />
+              <i class="fab fa-twitter"></i>
+            </a>
+          </li>
+        </div>
+      </ul>
+    </div>
+  </div>
+  <div class="footer-bottom">
     <div class="wrapper">
-        <p>Spotify© - 2020</p>
+      <p>Spotify &copy; - {{ getTodayDate() }}</p>
     </div>
-</div>
+  </div>
 </footer>
