+++ teste/Spotify/src/app/nav-bar/nav-bar.component.html
@@ -1,8 +1,21 @@
-<div class="header"> 
-    <ul> <!--Representa Lista de itens -->
-      <li><a href="index.html">Home</a></li>
-      <li><a href="#" target="_blank">Ajuda</a></li>
-      <li><a href="https://www.instagram.com/spotifybrasil/?hl=pt-br" target="_blank">Instagram</a></li> 
-      <li><a href="https://promo.spotify.com/br/store-1/?checkout=false&utm_expid=.GvcfMXEoRHSeQhQ9M5lIYA.1&utm_referrer=" target="_blank" >Premium</a></li>
-    </ul>
+<div class="header">
+  <ul>
+    <!--Representa Lista de itens -->
+    <li><a routerLink="">Home</a></li>
+    <li><a routerLink="/faq">Ajuda</a></li>
+    <li>
+      <a
+        href="https://www.instagram.com/spotifybrasil/?hl=pt-br"
+        target="_blank"
+        >Instagram</a
+      >
+    </li>
+    <li>
+      <a
+        href="https://promo.spotify.com/br/store-1/?checkout=false&utm_expid=.GvcfMXEoRHSeQhQ9M5lIYA.1&utm_referrer="
+        target="_blank"
+        >Premium</a
+      >
+    </li>
+  </ul>
 </div>
