+++ teste/Spotify/src/app/show-case/show-case.component.html
@@ -1,27 +1,25 @@
 <div class="circle"></div>
-  
-  <div class="logoAnimation">
-  <img src="assets/logoAnimation.gif" width="290" height="290" >
-  </div>
-  
-  <div class ="welcome">
-    <h1 > Bem-vindo ao <b>Spotify</b> </h1> 
-  </div>
 
-  
-  <div class="slogan">
-    <h2>Complete seu mundo com música</h2>
-  </div>
+<div class="logoAnimation">
+  <img src="assets/logoAnimation.gif" width="290" height="290" />
+</div>
 
+<div class="welcome">
+  <h1>Bem-vindo ao <b>Spotify</b></h1>
+</div>
 
-  <div class="WelcomeFamily">
-    <h1>Faça Parte de nossa familia</h1>
-    <h2>#NowFeeling</h2>
-  </div>
+<div class="slogan">
+  <h2>Complete seu mundo com música</h2>
+</div>
 
-  <div class="Comercial">
-    <video width="680" controls >
-      <source src="./assets/Comercial.mp4" type="video/mp4">
-    </video>
-  </div>
-  
+<div class="WelcomeFamily">
+  <h1>Faça Parte de nossa familia</h1>
+  <h2>#NowFeeling</h2>
+</div>
+
+<div class="Comercial">
+  <video width="680" controls>
+    <source src="./assets/Comercial.mp4" type="video/mp4" />
+  </video>
+</div>
+<app-footer></app-footer>
